+++ splc/splc.c
@@ -4,7 +4,7 @@ int main(int argc, const char **argv) {
    optflg_t of = {0};
    optval_t ov;
    array_t *toks;
-   int lc, wc;
+   size_t lc, wc;
 
    /* Initialization */
    init_msg();
